+++ src/server/Modules/Inmo/Modules.Inmo.Core/Features/Properties/Events/PropertyRegisteredEvent.cs
@@ -22,7 +22,7 @@ namespace InmoIT.Modules.Inmo.Core.Features.Properties.Events
 
         public string Description { get; set; }
 
-        public int SquareMeter { get; set; }
+        public decimal SquareMeter { get; set; }
 
         public int NumberRooms { get; set; }
